+++ lib/src/extensions/extensions.dart
@@ -29,7 +29,6 @@ import 'package:intl/intl.dart';
 
 import '../utils/constants/constants.dart';
 import '../utils/emoji_parser.dart';
-import '../utils/package_strings.dart';
 
 /// Extension for DateTime to get specific formats of dates and time.
 extension TimeDifference on DateTime {
@@ -46,9 +45,9 @@ extension TimeDifference on DateTime {
     final differenceInDays = currentDate.difference(targetDate).inDays;
 
     if (differenceInDays == 0) {
-      return PackageStrings.today;
+      return PackageStrings.currentLocale.today;
     } else if (differenceInDays <= 1 && differenceInDays >= -1) {
-      return PackageStrings.yesterday;
+      return PackageStrings.currentLocale.yesterday;
     } else {
       final DateFormat formatter = DateFormat(chatSeparatorDatePattern);
       return formatter.format(this);
@@ -130,11 +129,11 @@ extension ChatViewStateTitleExtension on String? {
       case ChatViewState.hasMessages:
         return this ?? '';
       case ChatViewState.noData:
-        return this ?? 'No Messages';
+        return this ?? PackageStrings.currentLocale.noMessage;
       case ChatViewState.loading:
         return this ?? '';
       case ChatViewState.error:
-        return this ?? 'Something went wrong !!';
+        return this ?? PackageStrings.currentLocale.somethingWentWrong;
     }
   }
 }